cleanup and document Color
